made separate calendar component and css
